# openapi.yaml
openapi: 3.0.3
info:
  title: Personal Investment Management API
  version: 0.2.0
  description: MVP API for personal investment management tool (snapshot-authoritative).
servers:
- url: '{{baseUrl}}'
  description: Variable base URL
components:
  schemas:
...
    ValuationPoint:
      type: object
      properties:
        product_id:
          type: integer
        date:
          type: string
          format: date
        market_value:
          type: number
      required:
      - product_id
      - date
      - market_value
...
  parameters:
    FromDate:
      name: from
      in: query
      schema:
        type: string
        format: date
      required: true
    ToDate:
      name: to
      in: query
      schema:
        type: string
        format: date
      required: true
...
paths:
...
  /api/valuations/batch_upsert:
    post:
      summary: Batch upsert product valuations
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      inserted:
                        type: integer
                      updated:
                        type: integer
                      warnings:
                        type: array
                        items:
                          type: string
                    required:
                    - inserted
                    - updated
                    - warnings
                  message:
                    type: string
                required:
                - code
                - data
                - message
              example:
                code: 200
                data:
                  inserted: 2
                  updated: 0
                  warnings: []
                message: ok
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rows:
                  type: array
                  items:
                    $ref: '#/components/schemas/ValuationPoint'
                source:
                  type: string
                  default: manual
              required:
              - rows
            example:
              rows:
              - product_id: 200
                date: '2025-01-05'
                market_value: 100059.46
              - product_id: 201
                date: '2025-01-05'
                market_value: 67747.76
  /api/products/{id}/valuations:
    get:
      summary: Get product valuations in range
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - $ref: '#/components/parameters/FromDate'
      - $ref: '#/components/parameters/ToDate'
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      product_id:
                        type: integer
                      points:
                        type: array
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                              format: date
                            market_value:
                              type: number
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          page_size:
                            type: integer
                          total:
                            type: integer
                          total_pages:
                            type: integer
                  message:
                    type: string
                required:
                - code
                - data
                - message
    delete:
      summary: Delete product valuation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: date
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                  message:
                    type: string
                required:
                - code
                - data
                - message
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                required:
                - code
                - message
...